pyproject.toml
[project]
name = "agentic-image-search"
version = "0.1.0"
description = "Starter scaffold for an agentic image search app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "sqlalchemy>=2.0.30",
  "chromadb>=0.5.0",
  "langchain>=0.2.0",
  "langgraph>=0.2.0",
  "python-dateutil>=2.9.0",
  "Pillow>=10.0.0",
  "transformers>=4.36.0",
  "opencv-python>=4.8.0",
  "openvino>=2024.0.0",
  "openvino-genai>=2024.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
